+++ htmlfmt.php
@@ -7,6 +7,7 @@ $LlCyfarwyddo="";
 $LlLlun1="";
 $LlLlun2="";
 $LlFfeil="";
+$LlFfeilPrev="";
 $LlLledYmylCy="";
 $LlLledYmylEn="";
 $LlMaintCy="";
@@ -26,7 +27,7 @@ $DFfeil1 = explode("\n", file_get_contents("./". $LlTestun) );
 //---------------------------------------
 //---------------------------------------
 function htmlfmtinit(){
-global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun;
+global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlFfeilPrev; global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun;
 $LlCwrs="";
 $LlGwers="";
 $LlTeitl="";
@@ -35,6 +36,7 @@ $LlCyfarwyddo="";
 $LlLlun1="";
 $LlLlun2="";
 $LlFfeil="";
+$LlFfeilPrev="";
 $LlLledYmylCy="";
 $LlLledYmylEn="";
 $LlMaintCy="";
@@ -46,7 +48,7 @@ $LlEn="";
 //---------------------------------------
 //---------------------------------------
 function htmlfmtsettings($ll1a){
-global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun;
+global $LlCwrs; global $LlGwers; global $LlTeitl; global $LlNodwch; global $LlCyfarwyddo; global $LlLlun1; global $LlLlun2; global $LlFfeil; global $LlFfeil;  global $LlLledYmylCy; global $LlLledYmylEn; global $LlMaintCy; global $LlMaintEn; global $LlCy; global $LlEn; global $DLlun1; global $DLlun2; global $DLlun3; global $DLlun4; global $DFfeil1; global $LlTestun;
 $bfound=false;
       //____$d1a = explode("=", substr($ll1a,1));
       $d1a = explode("=", substr($ll1a,0));
@@ -131,6 +133,7 @@ $bfound=false;
 
         $bfound=true;
       }else if($d1a[0]=="ffeil"){
+        $LlFfeilPrev=$LlFfeil;
         $LlFfeil=$d1a[1];
         $bfound=true;
  		 }